Refactoring the constants from a global scope to a more package-level scope to be more cohesion friendly
