+++ service/src/main/java/uk/nhs/adaptors/gp2gp/ehr/utils/EhrConstants.java
@@ -1,10 +1,10 @@
-package uk.nhs.adaptors.gp2gp.common.utils;
+package uk.nhs.adaptors.gp2gp.ehr.utils;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class AppConstants {
+public class EhrConstants {
 
     public static final String COMPLETE_CODE = "COMPLETE";
     public static final String CONTENT_TYPE_PLACEHOLDER = "CONTENT_TYPE_PLACEHOLDER_ID=";
